% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_noise.R
\name{add_noise}
\alias{add_noise}
\title{#########################################################################################
Function to  add noise to raw data.}
\usage{
add_noise(df, clustering.markers)
}
\arguments{
\item{df}{a dataframe/table of numeric values}

\item{clustering.markers}{an array of character strings defining the colname to use}
}
\value{
df a dataframe/table of numeric values
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
