% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_functions.R
\name{autoLgcl}
\alias{autoLgcl}
\title{#########################################################################################
Cytofkit's functions}
\source{
https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005112
}
\usage{
autoLgcl(x, channels, m = 4.5, q = 0.05)
}
\arguments{
\item{x}{a dataframe/table of numeric value}

\item{channels}{an array of character strings specifying the channels to use}

\item{m}{a numeric value}

\item{q}{a numeric value}

\item{value}{a dataframe/table of numeric value}

\item{cofactor}{a numeric value}
}
\value{
value a dataframe/table of numeric value
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
}
\details{
Adapted from :
Chen et al., 2016, PLOS computational biology
https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1005112
########################################################################################
}
