% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_directory.R
\name{choose_directory}
\alias{choose_directory}
\title{#########################################################################################
Function to choose a project directory.}
\usage{
choose_directory(caption = "Select a directory")
}
\arguments{
\item{caption}{a character string}
}
\value{
dir
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
