% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_FItSNE.dir.R
\name{choose_FItSNE.dir}
\alias{choose_FItSNE.dir}
\title{#########################################################################################
Function to choose FItSNE install directory.}
\usage{
choose_FItSNE.dir(dir)
}
\arguments{
\item{dir}{a character string specifying the project directory}
}
\value{
FItSNE.dir a character string containing the path to FItSNE folder
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
