% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_clusters.R
\docType{data}
\name{color_clusters}
\alias{color_clusters}
\title{#########################################################################################
Patrice Vallin's palette (55 colors)}
\format{An object of class \code{character} of length 56.}
\usage{
color_clusters
}
\value{
color_clusters an array of color in hex dec format
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
\keyword{datasets}
