% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colScale.R
\name{colScale}
\alias{colScale}
\title{#########################################################################################
Efficient scale function.}
\source{
https://www.r-bloggers.com/a-faster-scale-function/
}
\usage{
colScale(x, center = TRUE, scale = TRUE, add_attr = TRUE,
  rows = NULL, cols = NULL)
}
\arguments{
\item{x}{a table/dataframe}

\item{center}{a logical specifying if the data should be centered}

\item{scale}{a logical specifying if the data should be scaled}

\item{add_attr}{a logical}

\item{rows}{an array of numeric values}

\item{cols}{an array of numeric values}
}
\value{
x a table/dataframe
}
\description{
https://www.r-bloggers.com/a-faster-scale-function/
########################################################################################
}
