% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensional_reduction.R
\name{dimensional_reduction}
\alias{dimensional_reduction}
\title{Dimensionality reduction}
\usage{
dimensional_reduction(dir = NULL)
}
\arguments{
\item{dir}{a character string specifying the project directory}
}
\description{
Dimensionality reduction
}
