% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyFlow.analysis.R
\name{easyFlow.analysis}
\alias{easyFlow.analysis}
\title{Integrated and easy to run workFlow
for unsupervised data analysis.
raw data > preprocessing > clustering > dimensionality reduction > statistical analysis}
\usage{
easyFlow.analysis()
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
}
