% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyFlow.initialize.R
\name{easyFlow.initialize}
\alias{easyFlow.initialize}
\title{#########################################################################################
Initialize easyFlow}
\usage{
easyFlow.initialize()
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
