% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowAI.R
\name{flowai}
\alias{flowai}
\title{#########################################################################################
Compute a quality control of each samples and extract events in the specified gate of the metadata file}
\source{
http://bioconductor.org/packages/release/bioc/html/flowAI.html

https://academic.oup.com/bioinformatics/article/32/16/2473/2240408

https://www.bioconductor.org/packages/release/bioc/html/CytoML.html

https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.23663
}
\usage{
flowai(dir, files, metadata, exclude.markers.names, do_flowai,
  do_pregating)
}
\arguments{
\item{dir}{a character string specifying the project directory}

\item{files}{a array of character string specifying the fcs filenames}

\item{metadata}{a dataframe containing the metadata}

\item{exclude.markers.names}{a array of character string specifying the markers excluded of the analysis}

\item{do_flowai}{a logical specifying if the flowAI should be computed}

\item{do_pregating}{a logical specifying if the pregating should be computed}
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
