% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weber_helper_cluster_medians.R
\name{helper_cluster_medians}
\alias{helper_cluster_medians}
\title{#########################################################################################
Function to calculate median expression of each marker and scale to min = 0, max = 1,
for a given cluster. Note that data should already be transformed, e.g. using standard
asinh transform.}
\source{
https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.23030

https://github.com/lmweber/cytometry-clustering-comparison
}
\usage{
helper_cluster_medians(data, labels)
}
\arguments{
\item{data}{a table/dataframe}

\item{labels}{an array of numeric values}
}
\description{
Lukas Weber, July 2016
}
\details{
https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.23030
https://github.com/lmweber/cytometry-clustering-comparison
#########################################################################################
}
