% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_labels.R
\name{MEM_labels}
\alias{MEM_labels}
\title{#########################################################################################
Function to calculate MEM (Marker enrichment modeling) score.}
\usage{
MEM_labels(dir = NULL)
}
\arguments{
\item{dir}{a character string specifying the project directory}
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
