% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_report.R
\name{MEM_report}
\alias{MEM_report}
\title{#########################################################################################
Function to plot MEM report (Marker modeling enrichment).}
\usage{
MEM_report(dir = NULL)
}
\arguments{
\item{dir}{a character string specifying the project directory}
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
