% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaclustering_HC.R
\name{metaClustering_HC}
\alias{metaClustering_HC}
\title{Metaclustering using Hierarchical consensus clustering}
\source{
https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.22625

http://bioconductor.org/packages/release/bioc/html/FlowSOM.html
}
\usage{
metaClustering_HC(dir = NULL)
}
\arguments{
\item{dir}{a character string specifing the project directory}
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
SSE and findElbow function are adapted from FlowSOM R package
Sofie Van Gassen et al., Cytometry Part A, 2015
https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.22625
http://bioconductor.org/packages/release/bioc/html/FlowSOM.html
}
