% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacyto.R
\name{metacyto}
\alias{metacyto}
\title{#########################################################################################
Metacyto}
\source{
https://linkinghub.elsevier.com/retrieve/pii/S2211-1247(18)31080-5

https://bioconductor.org/packages/release/bioc/html/MetaCyto.html
}
\usage{
metacyto(dir = NULL)
}
\arguments{
\item{dir}{a character string specifying the project directory}
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
}
\details{
Adapted from
Hu et al., Cytometry Part A, 2018
https://linkinghub.elsevier.com/retrieve/pii/S2211-1247(18)31080-5
https://bioconductor.org/packages/release/bioc/html/MetaCyto.html
########################################################################################
}
