% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofworkflow_functions.R
\name{NRS}
\alias{NRS}
\title{#####################################################
CytofWorkflow's functions}
\source{
https://f1000research.com/articles/6-748

https://github.com/markrobinsonuzh/cytofWorkflow
}
\usage{
NRS(x, ncomp = 3)
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
}
\details{
Adapted from :
Nowicka M, Crowell H, Robinson M (2019).
cytofWorkflow: CyTOF workflow: differential discovery in high-throughput high-dimensional cytometry datasets.
https://f1000research.com/articles/6-748
https://github.com/markrobinsonuzh/cytofWorkflow.
#####################################################
}
