% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{#########################################################################################
Preprocessing data}
\usage{
preprocessing()
}
\value{
dir a character string specifying the project directory
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
