% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{#########################################################################################
Report function}
\usage{
report(dir = NULL)
}
\arguments{
\item{dir}{a character string specifying the project directory}
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
