% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_results.R
\name{save_results}
\alias{save_results}
\title{#########################################################################################
Save results functions}
\usage{
save_results(dir = NULL)
}
\arguments{
\item{dir}{a character string specifying the project directory}
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
