% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_metaclusters.R
\name{set_metaclusters}
\alias{set_metaclusters}
\title{Set metaclusters}
\usage{
set_metaclusters(dir = NULL)
}
\arguments{
\item{dir}{a character string specifing the project directory}
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
}
