% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/som_clustering.R
\name{som_clustering}
\alias{som_clustering}
\title{Clustering using self organizing map}
\usage{
som_clustering(dir = NULL)
}
\arguments{
\item{dir}{a character string specifing the project directory}
}
\description{
Patrice Vallin, easyFlow, Sept 2019.
}
