% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPADEvizR.R
\name{SPADEvizR}
\alias{SPADEvizR}
\title{#########################################################################################
SPADEVizR QC & statistical functions}
\source{
https://academic.oup.com/bioinformatics/article/33/5/779/2662307

https://github.com/tchitchek-lab/SPADEVizR
}
\usage{
SPADEvizR(dir = NULL)
}
\description{
Adapted from :
Guillaume Gautreau et al., Bioinformatics, March 2017
https://academic.oup.com/bioinformatics/article/33/5/779/2662307
https://github.com/tchitchek-lab/SPADEVizR
}
\details{
Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
