% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_analysis.R
\name{statistical_analysis}
\alias{statistical_analysis}
\title{#########################################################################################
Statistical tests functions and plots}
\source{
https://academic.oup.com/bioinformatics/article/33/5/779/2662307

https://github.com/tchitchek-lab/SPADEVizR

https://f1000research.com/articles/6-748

https://github.com/markrobinsonuzh/cytofWorkflow
}
\usage{
statistical_analysis(dir = NULL)
}
\arguments{
\item{dir}{a character string specifying the project directory}
}
\description{
Adapted from :
Guillaume Gautreau et al., Bioinformatics, March 2017
https://academic.oup.com/bioinformatics/article/33/5/779/2662307
https://github.com/tchitchek-lab/SPADEVizR
}
\details{
Adapted from :
Nowicka M, Crowell H, Robinson M (2019).
cytofWorkflow: CyTOF workflow: differential discovery in high-throughput high-dimensional cytometry datasets.
https://f1000research.com/articles/6-748
https://github.com/markrobinsonuzh/cytofWorkflow.

Patrice Vallin, easyFlow, Sept 2019.
########################################################################################
}
